abs() wants an integer

+++ src/gestalt.c
@@ -69,7 +69,7 @@ static void gestalt_float(t_gestalt *x, t_floatarg f)
 		if(pitch < 1) pitch = 0;
 		if(pitch > 127) pitch = 127;
 
-		interval = abs(pitch - x->x_lastpitch);
+		interval = abs(pitch - (int)x->x_lastpitch);
 		gestalt = (clock_gettimesince(x->x_lastontime)/x->x_reftime) + interval;
 
 		x->x_lastpitch = pitch;
